feat(scheduler): default to Calendar tab + Week view
Two UX changes:

- Schedulers nav lands on the Calendar tab, not List. /schedulers now
  redirects to /schedulers/calendar; the table view moves to
  /schedulers/list. Sidebar href stays at /schedulers so the entry
  highlights for both tabs.
- Calendar opens in Week view by default (was Month).
